endpoing adjustments, changed blur ammount, created single blanket, binary blanket css
