+++ db/cards/card_create.sql
@@ -1,3 +1,3 @@
-INSERT INTO cards (user_id, tree_id, parent_id)
+INSERT INTO cards (user_id, tree_type, parent_id)
 VALUES ( $1, $2, $3 )
 RETURNING *;
