+++ src/components/Menu/Menu.css
@@ -3,7 +3,7 @@
 	position: fixed;
 	bottom: 0;
 	left: 0;
-	z-index: 2;
+	z-index: 3;
 	transition: .5s;
 }
