+++ src/components/Products/BinaryBlanket/BinaryBlanket.css
@@ -1,0 +1,172 @@
+h1 {
+	font-size: 28px;
+	margin-top: 40px;
+	letter-spacing: 4px;
+	text-transform: uppercase;
+}
+
+h3{
+	width:60px;
+}
+
+input{
+	padding: 5px;
+	width:100%;
+	border: none;
+	border-radius: 3px;
+	/* border: 2px solid rgba(0,0,0,0.15); */
+	background: rgba(255,255,255,.5);
+}
+
+input:focus{background: rgba(255,255,255,1)}
+
+.binary-blanket-wrapper {
+	/* background: linear-gradient(to bottom, rgba(255,255,255,0.85) 0%,rgba(255,255,255,0.85) 100%), center url(https://images.unsplash.com/photo-1511497584788-876760111969?ixlib=rb-0.3.5&ixid=eyJhcHBfaWQiOjEyMDd9&s=bf3ecd34b3757e384e703b2e79243fb7&auto=format&fit=crop&w=2689&q=80); */
+	background-color: rgba(0,0,0,.1);
+}
+
+.binary-blanket{
+	display:flex;
+	flex-direction:row-reverse;
+	align-items: center;
+	justify-content: center;
+	height: 1170px;
+}
+
+.level1, .level2, .level3, .level4 {
+	margin:10px;
+	display:flex;
+	flex-direction:column;
+	justify-content:space-evenly;
+	width:20%;
+}
+
+.binary-pair{
+	display:flex;
+	flex-direction:column;
+	background-color:#A9BA9D;
+	/* border: 2px solid rgba(0,0,0,0.15); */
+	border-radius: 3px;
+	padding: 5px;
+	overflow: hidden;
+}
+
+.inner-level-1 {}
+.inner-level-2 {margin:180px 0;}
+.inner-level-3 {margin:60px 0;}
+.inner-level-4 {margin:3px 0;}
+
+.connector-wrapper-1, .connector-wrapper-2, .connector-wrapper-3 {
+	width:30px;
+	display:flex;
+	flex-direction:column;
+	align-items: center;
+	justify-content: space-evenly;	
+}
+
+.connector-wrapper-1 {height:1300px; justify-content: space-between;}
+.connector-wrapper-2 {height:1300px;}
+.connector-wrapper-3 {height:1300px;}
+
+
+.connector1-2, .connector2-3, .connector3-4 {
+	border-left:2px solid rgba(0,0,0,0.15);
+	border-top:2px solid rgba(0,0,0,0.15);
+	border-bottom:2px solid rgba(0,0,0,0.15);
+	width: 15px;
+}
+
+.connector1-2 {height:470px; margin-top:410px;}
+.connector2-3 {height:225px;}
+.connector3-4 {height:120px;}
+
+.person{
+	display:flex;
+	flex-direction:row;
+}
+
+.person h3{
+	margin: 5px;
+	letter-spacing: 3px;
+	font-size: 16px;
+	color: rgba(0,0,0,0.65);
+	text-transform: uppercase;
+	width: 100%;
+	text-align: center;
+}
+
+.name {
+	width:50%;
+	padding: 5px;
+	padding-right: 5px;
+}
+
+.dates{
+	display:flex;
+	flex-direction:column;
+	width:50%;
+	padding: 5px;
+	padding-left: 5px;
+}
+
+.date-inputs {
+	border-radius: 3px;
+	overflow: hidden;
+	/* border: 2px solid rgba(0,0,0,0.15); */
+}
+
+.date-inputs input {border-radius: 0px; border: 0px solid rgba(0,0,0,0.15);}
+
+.binary-header, .save-div {
+	display: flex;
+	justify-content: center;
+}
+
+.save-div button {
+	margin: 20px 0;
+	padding: 10px 60px;
+	color: #333;
+	border: 2px #333 solid;
+	background: transparent;
+	border-radius: 30px;
+	text-transform: uppercase;
+	font-weight: 700;
+	font-size: 16px;
+	letter-spacing: 2px;
+}
+
+.save-div button:hover {
+	color: white;
+	background: #333;
+}
+
+.saved-message-container {
+	height: 55px;
+	width: 100%;
+	position: fixed;
+	top: 75px;
+	left: 0;
+	box-shadow: 0px 5px 50px 0px rgba(0,0,0,0.1);
+	pointer-events: none;
+	opacity: 0;
+	transition: 1s;
+}
+
+.saved-message {
+	height: 55px;
+	width: 100%;
+	display: flex;
+	align-items: center;
+	justify-content: center;
+	color: #333;
+	user-select: none;
+}
+
+.saved-message-visible {
+	opacity: 1;
+}
+
+.saved-message a {
+	letter-spacing: 2px;
+	text-transform: uppercase;
+}
