+++ src/ducks/reducer.js
@@ -1,5 +1,5 @@
 const initialState = {
-  user_id      : null,
+  user_id      : 2,
   user_name    : null,
   user_email   : null,
   auth_id      : null,
@@ -23,7 +23,7 @@ export default function reducer(state = initialState, action) {
       user_email   : null,
       auth_id      : null,
       auth_picture : null,
-      user_admin   : null,
+      user_admin   : null
     })
     default : return state;
